Move Resolver to internal and add re-export module
